# data_analysis/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "Envelope_Chart_": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Chart"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "meta"
        ],
        "title": "Envelope[Chart]",
        "type": "object"
      },
...
      "Envelope_Page_City__": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Page_City_"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "meta"
        ],
        "title": "Envelope[Page[City]]",
        "type": "object"
      },
...
      "Envelope_dict_": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "meta"
        ],
        "title": "Envelope[dict]",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "只读访问已发布的统计批次。金额单位分、电量 Wh；日期 Asia/Shanghai，endDate exclusive。预测尚未实现。",
    "title": "Charging Analytics API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/cities": {
      "get": {
        "operationId": "listCities",
        "parameters": [
          {
            "in": "query",
            "name": "datasetId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Datasetid"
            }
          },
          {
            "in": "query",
            "name": "publishedBatchId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Publishedbatchid"
            }
          },
          {
            "in": "query",
            "name": "cityId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cityid"
            }
          },
          {
            "in": "query",
            "name": "stationId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stationid"
            }
          },
          {
            "description": "Asia/Shanghai business date, inclusive",
            "in": "query",
            "name": "startDate",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Asia/Shanghai business date, inclusive",
              "title": "Startdate"
            }
          },
          {
            "description": "Asia/Shanghai business date, exclusive",
            "in": "query",
            "name": "endDate",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Asia/Shanghai business date, exclusive",
              "title": "Enddate"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 1000000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Pagesize",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "required": false,
            "schema": {
              "default": "id",
              "maxLength": 32,
              "minLength": 1,
              "title": "Sortby",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sortOrder",
            "required": false,
            "schema": {
              "default": "asc",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Sortorder",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Page_City__"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Inconsistent filters"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Unknown dataset/city/station"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Published batch mismatch"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Invalid query/body"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Unexpected internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Published data or model unavailable"
          }
        },
        "summary": "Cities"
      }
    },
    "/api/v1/dashboard/charts": {
      "get": {
        "operationId": "dashboardCharts",
        "parameters": [
          {
            "in": "query",
            "name": "datasetId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Datasetid"
            }
          },
          {
            "in": "query",
            "name": "publishedBatchId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Publishedbatchid"
            }
          },
          {
            "in": "query",
            "name": "cityId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cityid"
            }
          },
          {
            "in": "query",
            "name": "stationId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stationid"
            }
          },
          {
            "description": "Asia/Shanghai business date, inclusive",
            "in": "query",
            "name": "startDate",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Asia/Shanghai business date, inclusive",
              "title": "Startdate"
            }
          },
          {
            "description": "Asia/Shanghai business date, exclusive",
            "in": "query",
            "name": "endDate",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Asia/Shanghai business date, exclusive",
              "title": "Enddate"
            }
          },
          {
            "in": "query",
            "name": "chart",
            "required": false,
            "schema": {
              "default": "energy",
              "enum": [
                "energy",
                "revenue",
                "utilization",
                "states",
                "service",
                "cohorts",
                "load"
              ],
              "title": "Chart",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "granularity",
            "required": false,
            "schema": {
              "default": "day",
              "enum": [
                "day",
                "hour"
              ],
              "title": "Granularity",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 400,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Chart_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Inconsistent filters"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Unknown dataset/city/station"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Published batch mismatch"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Invalid query/body"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Unexpected internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Published data or model unavailable"
          }
        },
        "summary": "Charts"
      }
    },
    "/api/v1/dashboard/overview": {
      "get": {
        "operationId": "dashboardOverview",
        "parameters": [
          {
            "in": "query",
            "name": "datasetId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Datasetid"
            }
          },
          {
            "in": "query",
            "name": "publishedBatchId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Publishedbatchid"
            }
          },
          {
            "in": "query",
            "name": "cityId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cityid"
            }
          },
          {
            "in": "query",
            "name": "stationId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stationid"
            }
          },
          {
            "description": "Asia/Shanghai business date, inclusive",
            "in": "query",
            "name": "startDate",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Asia/Shanghai business date, inclusive",
              "title": "Startdate"
            }
          },
          {
            "description": "Asia/Shanghai business date, exclusive",
            "in": "query",
            "name": "endDate",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Asia/Shanghai business date, exclusive",
              "title": "Enddate"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Overview_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Inconsistent filters"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Unknown dataset/city/station"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Published batch mismatch"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Invalid query/body"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Unexpected internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Published data or model unavailable"
          }
        },
        "summary": "Overview"
      }
    },
...
    "/api/v1/models": {
      "get": {
        "operationId": "modelCapabilities",
        "parameters": [
          {
            "in": "query",
            "name": "datasetId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Datasetid"
            }
          },
          {
            "in": "query",
            "name": "publishedBatchId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Publishedbatchid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ModelCapabilities_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Inconsistent filters"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Unknown dataset/city/station"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Published batch mismatch"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Invalid query/body"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Unexpected internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Published data or model unavailable"
          }
        },
        "summary": "Models"
      }
    },
...
    "/api/v1/predict/availability": {
      "post": {
        "operationId": "predictAvailabilityNotReady",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_PredictionResult_"
                }
              }
            },
            "description": "Future model-adapter success format. Not implemented: current valid requests only return 503 MODEL_NOT_READY."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Inconsistent filters"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Unknown dataset/city/station"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Published batch mismatch"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Invalid query/body"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_"
                }
              }
            },
            "description": "Unexpected internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ModelCapabilities_"
                }
              }
            },
            "description": "MODEL_NOT_READY: no trained model published"
          }
        },
        "summary": "Predict Availability"
      }
    },
...
  }
}